Splits send_response into set_status_code and set_response_body
Current HTTP middleware don't send a response immediately, or at least
might not. This is notably the case when using net/http over fasthttp.

This changes to set status code independent from the response body. A
future change will allow you to control whether calling "next" buffers
the response or not. This will be needed to do response scrubbing as
needed in coraza.

Signed-off-by: Adrian Cole <adrian@tetrate.io>