Removes fasthttp
This project is in proof-of-concept stage and only dapr had interest in
fasthttp. Since there are no other champions of it, and the model is
more difficult to intercept (pointers vs interfaces), this removes
fasthttp from the project.

Thanks @ItalyPaleAle for helping reduce the cost of experimentation!

See https://github.com/dapr/dapr/issues/4979

Signed-off-by: Adrian Cole <adrian@tetrate.io>