Changes get_x_body to reader based ABI (#21)
This changes g_x_body function to allow calling in a loop with a smaller
buffer than the stream size. In doing so, this allows guests to spool
requests larger than the input buffer to disk or otherwise. This doesn't
complete the full test suite, as that will occur after set_x_body is
switched to a writer based impl.

Signed-off-by: Adrian Cole <adrian@tetrate.io>