Consolidates HTTP header and body ABI
This reduces the burden to implement hosts by using the same logic for
request or response headers and trailers. Notably, this adds a `$kind`
parameter to header and host functions. This also removes the
single-entry header getter as it is simpler to only support and document
the multi-get ABI, particularly as the guest is passing in a mutable
buffer anyway (as it has to copy out as needed).

For the go-side of the implementation, I chose to keep the interface
broken out as it reduces the amount of changes to the code.

Signed-off-by: Adrian Cole <adrian@tetrate.io>