Reduces feature cardinality to 30
wit-bindgen defaults to representing flags as a sequence of i32, even
for the 64-bit case. To remain able to use flag markup, we need to
reduce the feature cardinality from 62 to 30. That said, there's little
impact because there aren't likely more than a dozen features we would
enable via flags. After that point, it feels more like config.

Signed-off-by: Adrian Cole <adrian@tetrate.io>