Makes log levels begin with zero
While not specified, component-model aka wit generate enums that begin
with the ordinal zero and there's no way to change that. This defers to
zero indexing vs the alternative which would be to use documentation to
declare which log levels map to which number. Remember, while enums are
logically opaque, the ordinals must match both host-guest and between
guests as there is no enum data type in wasm.

Signed-off-by: Adrian Cole <adrian@tetrate.io>