Switches module and export names to lower_snake_case (#33)
This follows existing naming conventions which prefer lower_snake_case
for module imports and function names. For example,
"wasi_snapshot_preview1" does not use hyphen-case. This is particularly
important for functions as hyphens are usually an unsupported character
in function names. This also corrects inconsistency in generated code
which depending on language would name parameters in different case, and
not hyphen-case.

Signed-off-by: Adrian Cole <adrian@tetrate.io>