Reverts change around module name allocation (#62)
This reverts to the former approach of external name management because
it ends up faster in practice with http-wasm.

The last commit used "anonymous modules" which is not a good fit for
http-wasm in how it works today. Basically, http-wasm is pooling modules
and these are re-used for sequential handler functions. The members of
the pool change at a very low rate, so there's not a lot of lock
contention on the module namespace. On the other hand, the normal path,
sequential performance is slower when we use anonymous modules.

Signed-off-by: Adrian Cole <adrian@tetrate.io>