tck: extracts client so we can test HTTP/2.0
Before, we couldn't control the TCK client, which is fine for HTTP/1.1.
However, in mosn, there's a different codepath for HTTP/1.1 vs HTTP/2.0.

This extracts the client for use in the TCK, so you can assign TLS
config as needed. This also addresses naming drift that happened between
the time we started and completed the ABI.

Signed-off-by: Adrian Cole <adrian@tetrate.io>